stm32/rcc: Move CtlHseClock() to common code.
Also check state of HSE before setting.

git-svn-id: svn+ssh://svn.code.sf.net/p/ethernut/code/trunk@6199 1ca430b3-8271-4a6b-9b0e-3dcac98a8924

+++ nut/include/arch/cm3/stm/stm32_clk.h
@@ -91,7 +91,7 @@ typedef enum
 extern uint32_t STM_ClockGet(int idx);
 
 /* Control functions for the separate clocks */
-extern int CtlHseClock( uint8_t ena);
+extern int CtlHseClock(int ena);
 extern int CtlHsiClock( uint8_t ena);
 extern int CtlPllClock( uint8_t ena);
